pyproject.toml
[project]
name = "webchat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "agno>=1.8",
    "llama-index-core>=0.12",
    "llama-index-vector-stores-qdrant>=0.4",
    "llama-index-embeddings-huggingface>=0.5",
    "qdrant-client>=1.12",
    "openai>=1.50",
    "trafilatura>=1.6",
    "streamlit>=1.40",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]
